perf: vectorized image augs and a torch.compile knob for HPT and Pi
The 300M HPT train step was launch-bound, not compute-bound: ~14,600 CUDA
kernel launches per step, 230 ms of CPU against 111 ms of GPU. Two things
dominated, and both are fixed here.

PerSampleAugs called torchvision's ColorJitter once per image so each sample
drew its own factors -- 64 Python round trips a step, each launching a few
dozen kernels and copying its factors from the host. BatchedColorJitter draws
the factors as one on-device tensor and runs each adjustment once over the
batch; a degenerate factor range reproduces torchvision's scalar adjustment
exactly (tests/unit/test_image_augs.py). Anything it cannot batch still goes
through the per-image loop.

model.compile.enabled=true routes to <algo>.compile_for_training(), which uses
nn.Module.compile rather than torch.compile(module) so state_dict keys survive
-- the wrapper form renames every parameter _orig_mod.* and would break
checkpoints written before the flag. That only routes __call__, so HPT compiles
the trunk, the flow head's denoiser and the image encoders (the stems and heads
are reached through named methods), and PI's call site moves to __call__.

On one H100 at batch 64 and the 360x640 frames mecka decodes to: HPT 245 ->
173 ms from the augs, -> 109 ms compiled (2.24x). Pi at batch 16: 827 -> 588 ms
compiled (1.41x). In fp32 the compiled loss is bit-identical to eager
(bench/check_compile_parity.py).

Also here: the flow head's attentions pass need_weights=False so they dispatch
to scaled_dot_product_attention; frames decode to float32 instead of numpy's
float64 default, halving the collate and H2D traffic (338 -> 169 MB a camera a
batch, bit-identical); a per-camera per-step device->host sync is gone; and the
batch is no longer deep-copied every step for an OT loss that is off.

bench/README.md records the harnesses, the numbers, what was tried and did not
pay (persistent_workers, pin_memory, more workers, max-autotune), and two loose
ends -- ModelWrapper never receives enable_grad_norm, and the shared emimic/
venv is off-pin for transformers.

+++ egomimic/rldb/zarr/zarr_dataset_multi.py
@@ -2231,7 +2231,18 @@ class ZarrDataset(torch.utils.data.Dataset):
                         idx = _next("JPEG decode failed", key=k)
                         retry = True
                         break
-                    data[k] = np.transpose(decoded, (2, 0, 1)) / 255.0
+                    # float32, not numpy's float64 default for `uint8 / float`:
+                    # process_batch_for_training casts to float32 on the GPU
+                    # anyway, so the extra 4 bytes a pixel buy nothing and
+                    # double the collate, the pin and the H2D copy. At 360x640
+                    # and batch 64 that is 338 MB a camera a batch instead of
+                    # 169. The result is bit-identical: there are only 256
+                    # possible values and float32(u)/float32(255) equals
+                    # float32(u / 255.0) for every one of them (see
+                    # tests/unit/test_image_decode_dtype.py).
+                    data[k] = np.transpose(decoded, (2, 0, 1)).astype(
+                        np.float32
+                    ) / np.float32(255.0)
                 elif zarr_key in self._json_keys:
                     if isinstance(data[k], np.ndarray):
                         data[k] = [self._decode_json_entry(v) for v in data[k]]
